app/requirements.txt
Flask==2.0.3
pdf2image==1.14.0
Pillow==9.1.0
